KAN 44 al KAN 62: En commits anteriores ya hice: - KAN 44 - Configuración del ecosistema Spring Boot y JPA (KANS 45, 46, 47, 48) - KAN 11- Repositories y Entities (KANS 49, 50, 51 ,52) - kAN 53 - Services, DTOs y Mappers  (KANS 54, 55, 56, 57)
Lo nuevo de este commit es:
- KAN 58 - Controllers, Endpoints y exception (KANS 59, 60, 61 ,62, 63)
- Y modificar los servicios para usar las excepciones y algunas validaciones.

+++ cafe_con_huellas/src/main/java/com/example/cafe_con_huellas/service/AdoptionDetailService.java
@@ -1,5 +1,7 @@
 package com.example.cafe_con_huellas.service;
 
+import com.example.cafe_con_huellas.exception.BadRequestException;
+import com.example.cafe_con_huellas.exception.ResourceNotFoundException;
 import com.example.cafe_con_huellas.model.entity.AdoptionDetail;
 import com.example.cafe_con_huellas.repository.AdoptionDetailRepository;
 import lombok.RequiredArgsConstructor;
@@ -23,16 +25,24 @@ public class AdoptionDetailService {
     // Buscar detalle por ID
     public AdoptionDetail findById(Long id) {
         return adoptionDetailRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Adoption detail not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Detalle de adopción no encontrado con ID: " + id));
     }
 
     // Guardar detalle
     public AdoptionDetail save(AdoptionDetail adoptionDetail) {
+        // Validación: Si ya existe un detalle para esta relación, no permitimos otro (1:1)
+        if (adoptionDetail.getId() == null &&
+                existsByRelationshipId(adoptionDetail.getRelationship().getId())) {
+            throw new BadRequestException("Esta relación ya tiene un registro de detalles de adopción.");
+        }
         return adoptionDetailRepository.save(adoptionDetail);
     }
 
     // Eliminar detalle
     public void deleteById(Long id) {
+        if (!adoptionDetailRepository.existsById(id)) {
+            throw new ResourceNotFoundException("No se puede eliminar. Registro no encontrado.");
+        }
         adoptionDetailRepository.deleteById(id);
     }
